refactor(common): move events/parsers package to common/parsers
- Move pkg/events/parsers to common/parsers
- Update import paths across pkg/, signatures/, and tests/
- Add replace directive in go.mod for local common module

The parsers package contains generic Linux syscall argument parsers
(clone flags, capabilities, socket types, etc.) with no Tracee-specific
logic, making it a perfect fit for the common module.
