+++ signatures/golang/anti_debugging_ptraceme.go
@@ -3,7 +3,7 @@ package main
 import (
 	"errors"
 
-	"github.com/aquasecurity/tracee/pkg/events/parsers"
+	"github.com/aquasecurity/tracee/common/parsers"
 	"github.com/aquasecurity/tracee/signatures/helpers"
 	"github.com/aquasecurity/tracee/types/detect"
 	"github.com/aquasecurity/tracee/types/protocol"
